EndGame finally works for players that are new and using an unused username

+++ GuessWhere/Models/LeaderBoard.cs
@@ -18,7 +18,7 @@ namespace GuessWhere.Models
         public int IDuser { get; set; }
         public int IDgame { get; set; }
         public string username { get; set; }
-        public double score { get; set; }
+        public decimal score { get; set; }
     
         public virtual Game Game { get; set; }
         public virtual User User { get; set; }
